pyproject.toml
[project]
name = "retinal-disease-classifier"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "timm>=1.0.0",
    "scikit-learn>=1.4.0",
    "pydantic-settings>=2.0.0",
    "pandas>=2.0.0",
    "pillow>=10.0.0",
    "numpy>=1.26.0",
]

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.ruff]
line-length = 100
